client gfx refactor (2): init and destory
